Add winget installation option to README
I recently submitted this app to the winget-pkgs repository, so it's now available to be installed from the winget package manager that's built into Windows. This PR would highlight this in the README so it's visible to more users. Please let me know if you release any updates and would like the winget package to be updated as well!